+++ tsconfig.json
@@ -55,8 +55,7 @@
       "@/generated/*": ["src/generated/*"],
       "@/hooks/*": ["src/hooks/*"],
       "@/utils/*": ["src/utils/*"],
-      "@/layouts/*": ["src/layouts/*"],
-      "@/reducers/*": ["src/reducers/*"]
+      "@/layouts/*": ["src/layouts/*"]
     } /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */,
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
